move javascript code into a src/ directory
