+++ package.json
@@ -8,16 +8,16 @@
     "url": "git+https://github.com/zelonewolf/openstreetmap-americana.git"
   },
   "scripts": {
-    "config": "cp configs/config.maptiler.js config.js",
-    "code_format": "npx prettier --write . && npx svgo -f icons/ --config configs/config.svgo.cjs",
+    "config": "cp src/configs/config.maptiler.js src/config.js",
+    "code_format": "npx prettier --write . && npx svgo -f icons/ --config src/configs/config.svgo.cjs",
     "clean": "rm -rf dist .parcel-cache",
     "presprites": "rm -rf dist/sprites",
     "sprites": "node scripts/sprites.js",
     "status_map": "node scripts/status_map.js",
     "taginfo": "node scripts/taginfo.js",
     "start": "run-s clean sprites serve",
-    "serve": "parcel index.html --open --port 1776",
-    "build:code": "parcel build index.html --public-url ./",
+    "serve": "parcel src/index.html --open --port 1776",
+    "build:code": "parcel build src/index.html --public-url ./",
     "build": "run-s clean sprites build:code taginfo status_map"
   },
   "dependencies": {
