+++ src/americana.js
