+++ src/configs/config.aws.js
