+++ src/configs/config.svgo.cjs
