+++ src/js/shield_canvas_draw.js
