+++ src/js/shield_defs.js
