+++ src/js/shield_text.js
