+++ src/layer/background.js
