+++ src/layer/highway_shield.js
